Statistically testing linear correlations
Statistically testing whether a linear correlation truly exists between two variables is much more easier in R. In our Jupyter notebook for this project, we made a heat map of correlations. Using this program, we want to checked whether there was statistical evidence that the market share of the largest carrier (`large_ms`) was in fact correlated to each of the other numerical variables from the heat map.